Get JSON working: current time entry

+++ TimeTracker/app/src/main/java/nz/ac/uclive/rog19/seng440/assignment1/components/TimeEntryListItem.kt
@@ -25,7 +25,7 @@ import java.time.format.FormatStyle
 fun TimeEntryListItem(
     modifier: Modifier = Modifier,
     timeEntry: TimeEntry,
-    projects: Map<Int, Project>,
+    projects: Map<Long, Project>,
     zoneId: ZoneId = Clock.systemDefaultZone().zone,
     now: Instant? = Instant.now()
 ) {
