+++ TimeTracker/app/src/main/java/nz/ac/uclive/rog19/seng440/assignment1/components/TimeEntryListView.kt
@@ -20,7 +20,7 @@ import java.time.ZoneId
 fun TimeEntryListView(
     modifier: Modifier = Modifier,
     entries: List<TimeEntry>,
-    projects: Map<Int, Project>,
+    projects: Map<Long, Project>,
     zoneId: ZoneId = Clock.systemDefaultZone().zone,
     now: State<Instant> = mutableStateOf(Instant.now())
 ) {
