+++ TimeTracker/app/src/main/java/nz/ac/uclive/rog19/seng440/assignment1/model/Project.kt
@@ -3,8 +3,8 @@ package nz.ac.uclive.rog19.seng440.assignment1.model
 import androidx.compose.ui.graphics.Color
 
 // https://api.track.toggl.com/api/v9/workspaces/{workspace_id}/projects
-class Project(id: Int, name: String, color: String) {
-    val id: Int = id
+class Project(id: Long, name: String, color: String) {
+    val id: Long = id
     val name: String = name
 
     /// Hex color code, with '#'
